Refactor Cython extensions:
* Spin off Keyset (ex ResultSet) in separate module
* Spin off SimpleGraph and Imr in separate module
* Change some LmdbTriplestore methods from cpdef to cdef
