+++ lakesuperior/api/resource.py
@@ -16,7 +16,7 @@ from lakesuperior.exceptions import (
 from lakesuperior import env, thread_env
 from lakesuperior.globals import RES_DELETED, RES_UPDATED
 from lakesuperior.model.ldp_factory import LDP_NR_TYPE, LdpFactory
-from lakesuperior.store.ldp_rs.lmdb_triplestore import SimpleGraph
+from lakesuperior.store.ldp_rs.graph import SimpleGraph
 
 
 logger = logging.getLogger(__name__)
