+++ tests/1_api/test_resource_api.py
@@ -14,7 +14,7 @@ from lakesuperior.exceptions import (
         TombstoneError)
 from lakesuperior.globals import RES_CREATED, RES_UPDATED
 from lakesuperior.model.ldpr import Ldpr
-from lakesuperior.store.ldp_rs.lmdb_triplestore import SimpleGraph, Imr
+from lakesuperior.store.ldp_rs.graph import SimpleGraph, Imr
 
 
 @pytest.fixture(scope='module')
